requirements.txt
requests==2.32.3
pillow==10.4.0
watchdog==4.0.1
semantic-version==2.10.0
psutil==6.0.0
